Added key post feature and further fleshed out flask skeleton.
Created notes blueprint that contains the following features:

	- A main index page that lists the current users notes.
	- The ability to create a new note.
	- The ability to remove a note.

Notes that are created will be added to the database then only be viewable
and deletable by the user who created them. Unauthorized users will be directed
to a 403 error page.

Created error handlers that take errors and redirect to custom error pages.
Currently errors 500, 404 and 403 are handled by this system.

Added CSS to stylize main navbar and user posts. Temporary fixes but make
overall look of the application a little easier on the eyes.