Implemented pagination feature on the index page.
Added pagination on the index, will currently display up to 5 posts
based on most recently posted notes and gives user ability to view older and
newer posts bared on navigation located on bottom.

Added view and update function on posts.

	- View displays the post title, author, modified and created date
	 as well as the options to delete and edit the post if you are the author.
         at the moment only the author can view a post but this will change when sharing is added.

	- Update function reuses the Noteform and new_notenote.html page but preloads the existing
	 post information into the fields for editing.

Created a validate_user decorator that can be used to ensure the note the user is trying to change exists
and that the user has the permission to do so. If the note does not exists then it throws a 404 error, if
the user is not the author the the file then they are unable to delete or edit the file.